Move NativeObject and RefCountedObject into GLib package; add Natives as utility class.
