+++ src/org/freedesktop/gstreamer/lowlevel/GValueAPI.java
@@ -32,6 +32,7 @@ import org.freedesktop.gstreamer.lowlevel.annotations.Invalidate;
 import com.sun.jna.Library;
 import com.sun.jna.NativeLong;
 import com.sun.jna.Pointer;
+import org.freedesktop.gstreamer.glib.NativeObject;
 
 /**
  *
