+++ src/org/freedesktop/gstreamer/lowlevel/GstTypes.java
@@ -25,6 +25,7 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import org.freedesktop.gstreamer.glib.NativeObject;
 
 /**
  *
