+++ src/org/freedesktop/gstreamer/lowlevel/MainLoop.java
@@ -29,6 +29,7 @@ import java.util.concurrent.FutureTask;
 import org.freedesktop.gstreamer.Gst;
 
 import com.sun.jna.Pointer;
+import org.freedesktop.gstreamer.glib.RefCountedObject;
 
 /**
  * The GLib main loop.
